os/filestore: fix result handling logic of destroy_collection
The problem here is if we fail to remove the original collection,
we may go on to removing the temp collection(if there is any) and
succeed(although the probability can be really small), then the
result code for the former operation can be overwritten, which is
unsafe and can be misleading.

This pr solves the above problem by starting the process of
parallelly removing a temp collection only if we succeed in
removing the original collection.

Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>