ceph.spec.in: disable lttng and babeltrace explicitly
on rhel7, we do not pacakge tracepoint probe shared libraries,
but "configure" script enables them if lttng is detected. and rpm
complains at seeing installed but not pacakged files. so we'd better
disable them explicity if we are not going to package them.
BTW, i failed to find the {lib}babeltrace-devel in rhel7 core. so
it's still disabled. and hence rbd-replay-prep is not included
in rhel7 as before.

Fixes: #14844
Signed-off-by: Kefu Chai <kchai@redhat.com>