mgr/dashboard: Push Grafana dashboards on startup
Because we didn't have viable way to populate Grafana dashboards in a
fully containerized context, we needed a different method. This will
work in packaged and containerized deployments.

On module startup, we attempt to push the dashboard JSON files to the
Grafana instance, handling cases where one isn't configured. We supply a
command (ceph dashboard update-grafana-dashboards) as well as an API
endpoint (/api/grafana/update_dashboards) as well.

Signed-off-by: Zack Cerza <zack@redhat.com>